rearrange new files location
