+++ web/src/App.tsx
@@ -2,9 +2,9 @@ import { KeycloakProvider } from '@react-keycloak/web';
 import { ThemeProvider } from 'emotion-theming';
 import React from 'react';
 import 'react-toastify/dist/ReactToastify.css';
-import keycloak from '../keycloak';
-import theme from '../theme';
 import AppRouter from './AppRouter';
+import keycloak from './keycloak';
+import theme from './theme';
 
 const App = () => {
   return (
