+++ web/src/index.tsx
@@ -1,7 +1,7 @@
 import { css, Global } from '@emotion/core';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App';
+import App from './App';
 import * as serviceWorker from './serviceWorker';
 import typography from './typography';
